[FLINK-39124] Update scopes and unnecessary longValue()

+++ flink-tests/src/test/java/org/apache/flink/test/checkpointing/StateCheckpointedITCase.java
@@ -131,7 +131,7 @@ class StateCheckpointedITCase extends StreamFaultToleranceTestBase {
 
         for (Map<Character, Long> map : ValidatingSink.maps) {
             for (Long count : map.values()) {
-                assertThat(count.longValue()).isEqualTo(NUM_STRINGS / 40L);
+                assertThat(count).isEqualTo(NUM_STRINGS / 40L);
             }
         }
     }
